Move Extents iterator to new iters::extents module
