+++ src/dir_htree.rs
@@ -10,8 +10,9 @@ use crate::dir_block::DirBlock;
 use crate::dir_entry::{DirEntry, DirEntryName};
 use crate::dir_entry_hash::dir_hash_md4_half;
 use crate::error::{Corrupt, Ext4Error, Incompatible};
-use crate::extent::{Extent, Extents};
+use crate::extent::Extent;
 use crate::inode::{Inode, InodeFlags, InodeIndex};
+use crate::iters::extents::Extents;
 use crate::iters::file_blocks::FileBlocks;
 use crate::path::PathBuf;
 use crate::util::{read_u16le, read_u32le, usize_from_u32};
