+++ src/iters.rs
@@ -6,5 +6,6 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+pub(crate) mod extents;
 pub(crate) mod file_blocks;
 pub(crate) mod read_dir;
